+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v2.13.0
+## New
+- Show total number of active filters on cocktails and ingredients page
+- Added approximate volume, calories and alcohol units to cocktail details
+- You can now save default bar units, user settings will still overwrite this
+- Added support for glass type volume
+
 # v2.12.0
 ## New
 - You can now filter cocktails by ignored ingredients